requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
langgraph==0.0.60
langchain==0.2.1
langchain-openai==0.1.7
pydantic==2.7.1
python-dateutil==2.9.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0